core/requirements.txt
python-decouple==3.1
scipy==1.2.0
georinex==1.7.1
coverage==4.5.2
dateparser==0.7.1